+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Removed
+- Removed the validator error factory interface from this library as the one provided by `cloudcreativity/json-api`
+has the additional methods on it.
+
 ## [0.8.0] - 2017-05-20
 
 ### Added